refactor: create exercise context to contain commonly used logic
