+++ src/pages/exercises.tsx
@@ -1,12 +1,9 @@
 import { NextPage } from 'next';
-import { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import ExerciseCard from '@components/ExerciseCard';
 import ExerciseForm from '@components/ExerciseForm';
+import Exercises from '@components/Exercises';
 import { protectedRoute } from '@utils/protectedRoute';
-import { supabase } from '@utils/supabaseClient';
-import { uppercase } from '@utils/uppercase';
-import { Exercise } from '../types';
+import { ExerciseContextProvider } from '@contexts/ExerciseContext';
 
 const StyledHeaderWrapper = styled.div`
   height: 200px;
@@ -28,110 +25,19 @@ const StyledH1 = styled.h1`
   padding-left: 10px;
 `;
 
-const StyledH2 = styled.h2`
-  text-align: center;
-  margin-bottom: 10px;
-`;
-
-const StyledH3 = styled.h3`
-  font-size: 1.2rem;
-  margin-bottom: 10px;
-`;
-
-const StyledMuscleGroupsWrapper = styled.div`
-  margin: 0 20px 20px 20px;
-`;
-
-const StyledExerciseWrapper = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(340px, 1fr));
-  flex-direction: column;
-  gap: 20px;
-`;
-
-const StyledP = styled.p`
-  padding: 0 20px 20px;
-  text-align: center;
-`;
-
 export const getServerSideProps = protectedRoute;
 
-const ExercisesPage: NextPage = () => {
-  const [exercises, setExercises] = useState<Exercise[]>([]);
-  const [synchronizeData, setSynchronizeData] = useState(false);
-
-  const getExercises = async () => {
-    try {
-      const user = supabase.auth.user();
-      const { data, error } = await supabase
-        .from('exercises')
-        .select()
-        .eq('userId', user?.id);
-
-      if (error) {
-        throw new Error(error.message);
-      }
-
-      setExercises(data);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    getExercises();
-  }, [synchronizeData]);
-
-  const muscleGroups = [
-    ...new Set(
-      exercises
-        .map((exercise) => exercise.muscleGroups)
-        .flat()
-        .map((muscleGroup) => muscleGroup.toLowerCase()),
-    ),
-  ].sort();
-
-  return (
+const ExercisesPage: NextPage = () => (
+  <ExerciseContextProvider>
     <>
       <StyledHeaderWrapper>
         <StyledHeaderImage />
         <StyledH1>Övningar</StyledH1>
       </StyledHeaderWrapper>
-
-      <ExerciseForm setSynchronizeData={setSynchronizeData} />
-
-      <>
-        <StyledH2>Dina övningar</StyledH2>
-        {muscleGroups.length === 0 ? (
-          <StyledP>
-            Du har inga övningar ännu men du kan lägga till en ovanför!
-            &#128170;
-          </StyledP>
-        ) : (
-          muscleGroups.map((muscleGroup) => (
-            <StyledMuscleGroupsWrapper key={muscleGroup}>
-              <StyledH3>{uppercase(muscleGroup)}</StyledH3>
-              <StyledExerciseWrapper>
-                {exercises
-                  .filter((exercise) =>
-                    exercise.muscleGroups
-                      .map((muscleGroup) => muscleGroup.toLowerCase())
-                      .includes(muscleGroup),
-                  )
-                  .map((exercise) => (
-                    <ExerciseCard
-                      key={exercise.id}
-                      exercise={exercise}
-                      setExercises={setExercises}
-                    />
-                  ))}
-              </StyledExerciseWrapper>
-            </StyledMuscleGroupsWrapper>
-          ))
-        )}
-      </>
+      <ExerciseForm />
+      <Exercises />
     </>
-  );
-};
+  </ExerciseContextProvider>
+);
 
 export default ExercisesPage;
